tempCompileBot/requirements.txt
python-telegram-bot==12.0.0
requests==2.22.0
firebase-admin
pytz